working and fixed versions

+++ cfg_editor_osx.sh
@@ -14,7 +14,7 @@
 where=$( pwd | rev | cut -c1-5 | rev )
 if [ "${where/arts}" = "$where" ]; then
 	echo "this should be in your parts folder"
-#	exit
+	exit
 fi
 
 
@@ -61,21 +61,11 @@ if [[ $input1 == "y" || $input1 == "Y" ]]; then
 			echo $file_location
 			echo
 			if [ -z "$mod_check" ]; then
-				# open the last bracket
-				tail -r $file_location > cfg_file.txt
-				#sed -i .bak 's/\}/\/\/\}/' cfg_file.txt
-				brack=$( cat $file_location | head -1 )
-				if [ -z $brack ]; then 
-					head="1d"
-					echo "no barcket"
-				else
-					brack=$( cat $file_location | head -2 )
-					if [ -z $brack ]; then
-						head="2d"
-						echo "no bracket2"
-					else
-						sed -i -e "$head" cfg_file.txt
-				tail -r cfg_file.txt > $file_location
+				# clean the extra lines
+				awk '/./' $file_location > cfg_file.txt
+				cfg_file="1"
+				# then cut the last line with the bracket
+				sed  '$ d' cfg_file.txt > $file_location
 				# then echo the module to the file, notice the extra bracket
 				echo "
 	MODULE {
@@ -96,8 +86,10 @@ if [[ $input1 == "y" || $input1 == "Y" ]]; then
 	done
 
 	echo
-	rm cfg_file.txt
-	#rm cfg_file.txt.bak
+	# make sure we only remove the file if its there
+	if [[ $cfg_file = "1" ]]; then
+		rm cfg_file.txt
+	fi
 	echo "done"
 else 
 	echo "okay"
